Add ceil_mode and custom kernel and stride support to max_pool_2d
My initial understanding of the MTCNN paper was flawwed: I had assumed the max
pooling was a typo and should have been 2x2 with the stride implicitly 2. In
fact, I was simply mistaken and it really is a 3x3 pool with a 2x2 stride.
Unfortunataely, for the sizes to work out a torch-style 'ceil_mode' behaviour
is required which adds significant complexity.