Replace einsum with tensordot in convolution implementation
This yields a >10x speedup for the Inception-ResNet-v1 network (though no change
to MTCNN).

The Numpy einsum optimiser implementation either has too high an overhead or
produces sufficiently poor results that it is not very performant in all cases.

The opt_einsum package offers a better optimiser which manages to make the
einsum implementation run in about 1/10th of the time. Since I don't want to
pull in an extra dependency I just looked at the optimisation plan which boils
down to performing the convolution as a tensor dot product.

It also finally forced me to grok what a tensor product is -- and that was
quite fun too. :)